clean version code
