+++ src/Styles/FavoritesPageStyles.js
@@ -1,0 +1,119 @@
+const styles = {
+    pageContainer: {
+      padding: '20px',
+      maxWidth: '1200px',
+      margin: '0 auto',
+    },
+    title: {
+      fontSize: '28px',
+      marginBottom: '24px',
+      textAlign: 'center',
+      color: '#333',
+    },
+    noFavorites: {
+      textAlign: 'center',
+      color: '#888',
+      fontSize: '16px',
+    },
+    cardGrid: {
+      display: 'grid',
+      gridTemplateColumns: 'repeat(auto-fit, minmax(280px, 1fr))',
+      gap: '50px 70px',
+      justifyContent: 'center',
+      marginTop: '40px',
+    },
+    card: {
+      borderRadius: '12px',
+      boxShadow: '0 4px 12px rgba(0, 0, 0, 0.1)',
+      padding: '20px',
+      backgroundColor: '#fff',
+      width: '280px',
+      display: 'flex',
+      flexDirection: 'column',
+      justifyContent: 'space-between',
+      height: '100%',
+    },
+    imageSection: {
+      display: 'flex',
+      gap: '10px',
+      marginTop: '10px',
+    },
+    largeImage: {
+      width: '140px',
+      height: '140px',
+      borderRadius: '8px',
+      objectFit: 'cover',
+      boxShadow: '0 2px 6px rgba(0,0,0,0.2)',
+      transition: 'transform 0.2s',
+    },
+    moreOverlay: {
+      position: 'absolute',
+      top: '0',
+      left: '0',
+      width: '100%',
+      height: '100%',
+      backgroundColor: 'rgba(0,0,0,0.4)',
+      color: '#fff',
+      fontWeight: 'bold',
+      fontSize: '18px',
+      display: 'flex',
+      justifyContent: 'center',
+      alignItems: 'center',
+      borderRadius: '8px',
+    },
+    buttonRow: {
+      marginTop: 'auto',
+      display: 'flex',
+      justifyContent: 'flex-start',
+      gap: '10px',
+    },
+    removeButton: {
+      padding: '5px 8px',
+      borderRadius: '6px',
+      border: 'none',
+      backgroundColor: '#f44336',
+      color: '#fff',
+      fontWeight: 'bold',
+      fontSize: '13px',
+      cursor: 'pointer',
+    },
+    connectButton: {
+      padding: '5px 8px',
+      borderRadius: '6px',
+      border: 'none',
+      fontWeight: 'bold',
+      fontSize: '13px',
+      color: '#fff',
+    },
+    modalOverlay: {
+      position: 'fixed',
+      top: 0,
+      left: 0,
+      right: 0,
+      bottom: 0,
+      backgroundColor: 'rgba(0,0,0,0.6)',
+      display: 'flex',
+      alignItems: 'center',
+      justifyContent: 'center',
+      zIndex: 999,
+    },
+    modalContent: {
+      backgroundColor: '#fff',
+      padding: '20px',
+      borderRadius: '12px',
+      maxWidth: '90%',
+      maxHeight: '80%',
+      overflowY: 'auto',
+      display: 'flex',
+      flexWrap: 'wrap',
+      gap: '10px',
+    },
+    modalImage: {
+      maxWidth: '200px',
+      maxHeight: '200px',
+      objectFit: 'cover',
+      borderRadius: '8px',
+    },
+  };
+
+  export default styles;
